#129 fix: 라우터 변경으로 인한 응답 데이터 변경 적용, 리뷰 등록, 삭제, 화면 렌더 로직 구연 완료

+++ src/static/js/restaurantDetail.js
@@ -21,45 +21,53 @@ window.addEventListener('resize', checkWidth);
 // 찜
 document.querySelector('.heart').addEventListener('click', async e => {
   const currentSrc = e.target.getAttribute('src');
+  if (userInfo) {
+    if (currentSrc === '/static/img/heart.png') {
+      try {
+        const response = await axios({
+          method: 'POST',
+          url: `/api/like`,
+          data: { restaurant_id: restaurant.restaurant_id },
+        });
 
-  if (currentSrc === '/static/img/heart.png') {
-    try {
-      const response = await axios({
-        method: 'POST',
-        url: `/api/like`,
-        data: { restaurant_id: restaurant.restaurant_id },
-      });
-
-      if (response.status === 201) {
-        alert('찜 목록에 추가되었습니다.');
-        e.target.setAttribute('src', '/static/img/heartFilled.png');
+        if (response.status === 201) {
+          alert('찜 목록에 추가되었습니다.');
+          e.target.setAttribute('src', '/static/img/heartFilled.png');
+        }
+      } catch (error) {
+        console.error('에러 정보:', error);
+        alert('찜 목록에 추가하는 것을 실패했습니다.');
       }
-    } catch (error) {
-      console.error('에러 정보:', error);
-      alert('로그인 후 찜이 가능합니다.');
-    }
-  } else if (currentSrc === '/static/img/heartFilled.png') {
-    try {
-      const response = await axios({
-        method: 'DELETE',
-        url: `/api/like`,
-        data: { restaurant_id: restaurant.restaurant_id },
-      });
+    } else if (currentSrc === '/static/img/heartFilled.png') {
+      try {
+        const response = await axios({
+          method: 'DELETE',
+          url: `/api/like`,
+          data: { restaurant_id: restaurant.restaurant_id },
+        });
 
-      if (response.status === 200) {
-        alert('찜 목록에서 삭제되었습니다.');
-        e.target.setAttribute('src', '/static/img/heart.png');
+        if (response.status === 200) {
+          alert('찜 목록에서 삭제되었습니다.');
+          e.target.setAttribute('src', '/static/img/heart.png');
+        }
+      } catch (error) {
+        console.error('에러 정보:', error);
+        alert('찜 목록에서 삭제하는 것을 실패했습니다.');
       }
-    } catch (error) {
-      console.error('에러 정보:', error);
-      alert('찜 목록에서 삭제하는 것을 실패했습니다.');
     }
+  } else {
+    alert('로그인 후 식당 찜이 가능합니다.');
   }
 });
 
 // 전화 걸기 or 전화번호 복사
 callButton.addEventListener('click', function () {
   const phoneNumber = restaurant.restaurant_phone;
+
+  if (!phoneNumber || phoneNumber.trim() === '') {
+    alert('전화번호가 등록되어 있지 않습니다.');
+    return;
+  }
 
   if (
     /Android|webOS|iPhone|iPad|iPod|BlackBerry|IEMobile|Opera Mini/i.test(
@@ -83,6 +91,10 @@ callButton.addEventListener('click', function () {
 document
   .getElementById('showReviewForm')
   .addEventListener('click', function () {
+    if (!userInfo) {
+      alert('로그인 후 리뷰 작성이 가능합니다.');
+      return;
+    }
     if (
       reviewForm.style.display === 'none' ||
       reviewForm.style.display === ''
@@ -136,14 +148,7 @@ document.getElementById('submitbtn').addEventListener('click', async e => {
     }
   } catch (error) {
     console.error('에러 정보:', error);
-
-    if (
-      error.response.data.message === '세션에서 사용자 정보를 찾을 수 없습니다.'
-    ) {
-      alert('로그인 후 리뷰를 등록해주세요.');
-    } else {
-      alert('리뷰를 등록하는 동안 오류가 발생했습니다.');
-    }
+    alert('리뷰를 등록하는 동안 오류가 발생했습니다.');
   }
 });
 
@@ -167,15 +172,11 @@ const addNewReview = reviewData => {
   newReviewElement.innerHTML = `
   <div>
     <div>
-     <p>작성자</p>
+     <p>${userInfo.user_name}</p>
     </div>
      <div>
-     <p><img src="/static/img/star.png" alt="평점" class="starIcon"> ${
-       reviewData.rating
-     }</p>
-    <p><img src="/static/img/like.png" alt="추천하기" class="thumbsIcon" id=${
-      reviewData.review_id
-    }> ${reviewData.is_useful === undefined ? 0 : reviewData.is_useful}</p>
+     <p><img src="/static/img/star.png" alt="평점" class="starIcon"> ${reviewData.rating}</p>
+     <p><img src="/static/img/like.png" alt="추천하기" class="thumbsIcon" id=${review.review_id}> <span>0<span></p>
      <p>${formattedDate}</p>
     </div>
     </div>
@@ -186,105 +187,381 @@ const addNewReview = reviewData => {
   renderReviews.prepend(newReviewElement);
 };
 
-// 지도
-let geocoder;
-let map;
-
-function initMap() {
-  geocoder = new google.maps.Geocoder();
-  const latlng = new google.maps.LatLng(-34.397, 150.644);
-  const mapOptions = {
-    zoom: 17,
-    center: latlng,
-  };
-
-  map = new google.maps.Map(document.querySelector('#map'), mapOptions);
-  codeAddress();
-}
-
-function codeAddress() {
-  const address = restaurant.restaurant_address;
-
-  geocoder.geocode({ address: address }, function (results, status) {
-    if (status == 'OK') {
-      map.setCenter(results[0].geometry.location);
-      const marker = new google.maps.Marker({
-        map: map,
-        position: results[0].geometry.location,
-      });
-    } else {
-      alert('Geocode was not successful for the following reason: ' + status);
-    }
-  });
-}
-
 // 리뷰 페이지네이션
 let currentPage = 1;
 const reviewsPerPage = 5;
 
-const renderReviewsPage = page => {
-  const start = (page - 1) * reviewsPerPage;
-  const end = page * reviewsPerPage;
+// const renderReviewsPage = page => {
+//   const start = (page - 1) * reviewsPerPage;
+//   const end = page * reviewsPerPage;
 
-  const reviews = restaurant.Reviews.reverse();
+//   const reviews = review.reverse();
 
-  const reviewsToRender = reviews.slice(start, end);
-  const reviewsElement = document.querySelector('.renderReviews');
+//   const reviewsToRender = reviews.slice(start, end);
+//   const reviewsElement = document.querySelector('.renderReviews');
 
-  reviewsElement.innerHTML = ''; // 리뷰 컨테이너 초기화
+//   reviewsElement.innerHTML = ''; // 리뷰 컨테이너 초기화
 
-  if (reviewsToRender.length === 0) {
-    reviewsElement.innerHTML = '<p>후기가 없습니다.</p>';
-    return;
-  }
+//   if (reviewsToRender.length === 0) {
+//     reviewsElement.innerHTML = '<p>후기가 없습니다.</p>';
+//     return;
+//   }
 
-  reviewsToRender.forEach(review => {
-    let imagesHTML = '';
+//   reviewsToRender.forEach(review => {
+//     let imagesHTML = '';
 
-    review.ReviewImages.forEach(image => {
-      imagesHTML += `<img src="${image.image_url}" alt="후기 이미지">`;
-    });
+//     review.ReviewImages.forEach(image => {
+//       imagesHTML += `<img src="${image.image_url}" alt="후기 이미지">`;
+//     });
 
-    console.log(review.reviewUsefulness);
+//     console.log(review.reviewUsefulness);
 
-    let imgSrc;
-    review.reviewUsefulness.length > 0
-      ? (imgSrc = '/static/img/likeFilled.png')
-      : (imgSrc = '/static/img/like.png');
+//     let imgSrc;
+//     if (userInfo) {
+//       if (
+//         review.reviewUsefulness.length > 0 &&
+//         review.reviewUsefulness.some(
+//           usefulness => usefulness.user_id === userInfo.id,
+//         )
+//       ) {
+//         imgSrc = '/static/img/likeFilled.png';
+//       } else {
+//         imgSrc = '/static/img/like.png';
+//       }
+//     } else {
+//       imgSrc = '/static/img/like.png';
+//     }
 
-    const reviewHTML = `
-    <section class="reviewContainer" >
-    <div>
+//     const reviewHTML = `
+//     <section class="reviewContainer" id="${review.review_id}">
+//     <div>
+//       <div>
+//        <p>${review.user.user_name}</p>
+//       </div>
+//        <div>
+//        <p class="ratingContainer"><img src="/static/img/star.png" alt="평점" class="starIcon"> <span class="reviewRatingText">${
+//          review.rating
+//        }</span></p>
+//       <p><img src=${imgSrc} alt="추천하기" class="thumbsIcon" id=${
+//       review.review_id
+//     }> <span>${
+//       review.reviewUsefulness.length > 0 ? review.reviewUsefulness.length : 0
+//     }<span></p>
+//        <p>${new Date(review.updatedAt)
+//          .toLocaleDateString()
+//          .replace(/\.$/, '')}</p>
+//       </div>
+//       </div>
+//       <div class="reviewImg">${imagesHTML}</div>
+//       <p class="reviewContent">${review.content}</p>
+//       <div class="editNdelete">
+//       <input id="edit" type="button" value="수정" />
+//       <input id="delete" type="button" value="삭제" />
+//     </section>
+//     `;
+
+//     reviewsElement.innerHTML += reviewHTML;
+//   });
+// };
+
+// const reviewsElement = document.querySelector('.renderReviews');
+
+// // 리뷰 컨테이너에 클릭 이벤트 리스너를 등록
+// reviewsElement.addEventListener('click', async e => {
+//   // "수정" 버튼이 클릭되었는지 확인
+//   if (e.target.id === 'edit') {
+//     // 수정 버튼을 포함하고 있는 가장 가까운 리뷰 컨테이너를 찾음
+//     const reviewContainer = e.target.closest('.reviewContainer');
+
+//     const reviewRatingElement =
+//       reviewContainer.querySelector('.reviewRatingText');
+
+//     // 해당 리뷰 컨테이너 내의 리뷰 내용을 담은 p 태그를 선택
+//     const reviewContentElement =
+//       reviewContainer.querySelector('.reviewContent');
+
+//     const editSubmit = reviewContainer.querySelector('#edit');
+
+//     // 원래의 리뷰 내용을 저장
+//     const originalContent = reviewContentElement.innerText;
+
+//     // 원래의 리뷰 평점을 저장
+//     const originalRating = parseFloat(reviewRatingElement.innerText);
+
+//     // 원래의 내용을 가진 textarea 요소를 생성
+//     const textareaElement = document.createElement('textarea');
+//     textareaElement.value = originalContent;
+//     textareaElement.classList.add('reviewContentTextarea');
+
+//     // 원래의 평점을 가진 select 요소를 생성
+//     const selectElement = document.createElement('select');
+//     for (let i = 5; i >= 0.5; i -= 0.5) {
+//       const optionElement = document.createElement('option');
+//       optionElement.value = i;
+//       optionElement.innerText = i;
+//       if (i === originalRating) {
+//         optionElement.selected = true;
+//       }
+//       selectElement.appendChild(optionElement);
+//     }
+//     selectElement.classList.add('reviewRatingSelect');
+
+//     // p 태그를 textarea로 교체
+//     reviewContentElement.replaceWith(textareaElement);
+//     reviewRatingElement.replaceWith(selectElement);
+
+//     // 선택적으로, textarea에 포커스를 줘서 사용자가 바로 입력할 수 있도록 함
+//     textareaElement.focus();
+//     editSubmit.value = '수정 완료';
+
+//     editSubmit.addEventListener('click', async () => {
+//       try {
+//         const response = await axios({
+//           method: 'PATCH',
+//           url: `/api/review/myreview/${reviewContainer.id}`,
+//           data: {
+//             content: textareaElement.value,
+//             rating: selectElement.value,
+//           },
+//         });
+
+//         if (response.status === 200) {
+//           alert('해당 리뷰가 수정되었습니다.');
+//         }
+
+//         // 수정된 리뷰 내용으로 p 태그를 생성
+//         const newContentElement = document.createElement('p');
+//         newContentElement.classList.add('reviewContent');
+//         newContentElement.innerText = textareaElement.value;
+
+//         // 수정된 리뷰 평점으로 p 태그를 생성
+//         const ratingContainer = document.querySelector('.ratingContainer');
+//         const newRatingElement = document.createElement('span');
+//         newRatingElement.innerHTML = selectElement.value;
+//         ratingContainer.append(newRatingElement);
+
+//         // textarea와 select를 새로운 p 태그로 교체
+//         textareaElement.replaceWith(newContentElement);
+//         selectElement.replaceWith(newRatingElement);
+
+//         // 수정 완료 버튼의 텍스트를 원래대로 되돌림
+//         editSubmit.value = '수정';
+//       } catch (error) {
+//         console.error('에러 정보:', error);
+//         alert('리뷰 수정을 실패했습니다.');
+//       }
+//     });
+//   }
+
+//   if (e.target.id === 'delete') {
+//     const reviewContainer = e.target.closest('.reviewContainer');
+//     alert(reviewContainer.id);
+//     try {
+//       const response = await axios({
+//         method: 'DELETE',
+//         url: `/api/review/myreview/${reviewContainer.id}`,
+//         data: { review_id: reviewContainer.id },
+//       });
+
+//       if (response.status === 200) {
+//         alert('해당 리뷰가 삭제되었습니다.');
+//       }
+//     } catch (error) {
+//       console.error('에러 정보:', error);
+//       alert('리뷰 삭제를 실패했습니다.');
+//     }
+//   }
+// });
+
+// renderReviewsPage(currentPage);
+
+const renderReview = review => {
+  const imagesHTML = review.ReviewImages
+    ? review.ReviewImages.map(
+        image => `<img src="${image.image_url}" alt="후기 이미지">`,
+      ).join('')
+    : '';
+
+  const imgSrc =
+    userInfo && review.reviewUsefulness.some(u => u.user_id === userInfo.id)
+      ? '/static/img/likeFilled.png'
+      : '/static/img/like.png';
+
+  const usefulnessCount = review.reviewUsefulness
+    ? review.reviewUsefulness.length
+    : 0;
+
+  // Edit and delete buttons are shown only if the logged-in user is the author of the review
+  const editDeleteHTML =
+    userInfo && userInfo.user_name === review.user.user_name
+      ? `
+    <div class="editNdelete">
+      <input id="edit" type="button" value="수정" />
+      <input id="delete" type="button" value="삭제" />
+    </div>`
+      : '';
+
+  return `
+    <section class="reviewContainer" id="${review.review_id}">
       <div>
-       <p>${review.user.user_name}</p>
-      </div>
-       <div>
-       <p><img src="/static/img/star.png" alt="평점" class="starIcon"> ${
-         review.rating
-       }</p>
-      <p><img src=${imgSrc} alt="추천하기" class="thumbsIcon" id=${
-      review.review_id
-    }> <span>${
-      review.reviewUsefulness.length > 0 ? review.reviewUsefulness.length : 0
-    }<span></p>
-       <p>${new Date(review.updatedAt)
-         .toLocaleDateString()
-         .replace(/\.$/, '')}</p>
-      </div>
+        <div>
+          <p>${review.user.user_name}</p>
+        </div>
+        <div>
+          <p class="ratingContainer"><img src="/static/img/star.png" alt="평점" class="starIcon"> <span class="reviewRatingText">${
+            review.rating
+          }</span></p>
+          <p><img src=${imgSrc} alt="추천하기" class="thumbsIcon" id=${
+    review.review_id
+  }> <span>${usefulnessCount}<span></p>
+          <p>${new Date(review.updatedAt)
+            .toLocaleDateString()
+            .replace(/\.$/, '')}</p>
+        </div>
       </div>
       <div class="reviewImg">${imagesHTML}</div>
       <p class="reviewContent">${review.content}</p>
+      ${editDeleteHTML}
     </section>
-    `;
-
-    reviewsElement.innerHTML += reviewHTML;
-  });
+  `;
 };
+
+const replaceWithEditableFields = (
+  reviewContainer,
+  originalContent,
+  originalRating,
+) => {
+  const textareaElement = document.createElement('textarea');
+  const selectElement = document.createElement('select');
+  textareaElement.value = originalContent;
+
+  selectElement.innerHTML = Array.from(
+    { length: 10 },
+    (_, i) => `<option value="${(i + 1) / 2}">${(i + 1) / 2}</option>`,
+  ).join('');
+  selectElement.value = originalRating;
+
+  const reviewContentElement = reviewContainer.querySelector('.reviewContent');
+  const reviewRatingElement =
+    reviewContainer.querySelector('.reviewRatingText');
+
+  if (reviewContentElement) {
+    reviewContentElement.replaceWith(textareaElement);
+  }
+
+  if (reviewRatingElement) {
+    reviewRatingElement.replaceWith(selectElement);
+  }
+
+  return { textareaElement, selectElement };
+};
+
+const replaceWithOriginalFields = (
+  textareaElement,
+  selectElement,
+  newContent,
+  newRating,
+) => {
+  const newContentElement = document.createElement('p');
+  const newRatingElement = document.createElement('span');
+  newContentElement.innerText = newContent;
+  newRatingElement.innerText = newRating;
+
+  textareaElement.replaceWith(newContentElement);
+  selectElement.replaceWith(newRatingElement);
+};
+
+const handleEditClick = async reviewContainer => {
+  const reviewContentElement = reviewContainer.querySelector('.reviewContent');
+  const reviewRatingElement =
+    reviewContainer.querySelector('.reviewRatingText');
+  const editSubmit = reviewContainer.querySelector('#edit');
+
+  const originalContent = reviewContentElement
+    ? reviewContentElement.innerText
+    : '';
+  const originalRating = reviewRatingElement
+    ? parseFloat(reviewRatingElement.innerText)
+    : '';
+
+  const { textareaElement, selectElement } = replaceWithEditableFields(
+    reviewContainer,
+    originalContent,
+    originalRating,
+  );
+
+  editSubmit.value = '수정 완료';
+
+  if (editSubmit.value === '수정 완료') {
+    editSubmit.addEventListener('click', async () => {
+      try {
+        const response = await axios({
+          method: 'PATCH',
+          url: `/api/review/myreview/${reviewContainer.id}`,
+          data: {
+            content: textareaElement.value,
+            rating: selectElement.value,
+          },
+        });
+
+        if (response.status === 200) {
+          alert('해당 리뷰가 수정되었습니다.');
+          replaceWithOriginalFields(
+            textareaElement,
+            selectElement,
+            textareaElement.value,
+            selectElement.value,
+          );
+          editSubmit.value = '수정';
+        }
+      } catch (error) {
+        console.error('에러 정보:', error);
+        alert('리뷰 수정을 실패했습니다.');
+      }
+    });
+  }
+};
+
+const handleDeleteClick = async reviewContainer => {
+  try {
+    const response = await axios({
+      method: 'DELETE',
+      url: `/api/review/myreview/${reviewContainer.id}`,
+      data: { review_id: reviewContainer.id },
+    });
+
+    if (response.status === 200) {
+      alert('해당 리뷰가 삭제되었습니다.');
+      window.location.reload();
+    }
+  } catch (error) {
+    console.error('에러 정보:', error);
+    alert('리뷰 삭제를 실패했습니다.');
+  }
+};
+
+const renderReviewsPage = page => {
+  const start = (page - 1) * reviewsPerPage;
+  const end = page * reviewsPerPage;
+  const reviewsToRender = review.slice(start, end);
+  const reviewsElement = document.querySelector('.renderReviews');
+  reviewsElement.innerHTML = reviewsToRender.length
+    ? reviewsToRender.map(review => renderReview(review, userInfo)).join('')
+    : '<p>후기가 없습니다.</p>';
+};
+
+const reviewsElement = document.querySelector('.renderReviews');
+
+reviewsElement.addEventListener('click', async e => {
+  const reviewContainer = e.target.closest('.reviewContainer');
+  if (e.target.id === 'edit') handleEditClick(reviewContainer, userInfo);
+  if (e.target.id === 'delete') handleDeleteClick(reviewContainer);
+});
 
 renderReviewsPage(currentPage);
 
 document.querySelector('#nextPage').addEventListener('click', () => {
-  if (currentPage * reviewsPerPage < restaurant.Reviews.length) {
+  if (currentPage * reviewsPerPage < review.length) {
     currentPage++;
     renderReviewsPage(currentPage);
   } else {
@@ -308,12 +585,14 @@ document
     const target = e.target;
 
     if (target.classList.contains('thumbsIcon')) {
+      if (!userInfo) {
+        alert('로그인 후 리뷰를 추천해주세요.');
+        return;
+      }
       const review_id = target.id;
       const currentSrc = target.getAttribute('src');
 
-      if (id === 0) {
-        alert('로그인 후 찜이 가능합니다.');
-      } else if (currentSrc === '/static/img/like.png') {
+      if (currentSrc === '/static/img/like.png') {
         try {
           const response = await axios({
             method: 'POST',
@@ -358,3 +637,35 @@ document
       }
     }
   });
+
+// 지도
+let geocoder;
+let map;
+
+function initMap() {
+  geocoder = new google.maps.Geocoder();
+  const latlng = new google.maps.LatLng(-34.397, 150.644);
+  const mapOptions = {
+    zoom: 17,
+    center: latlng,
+  };
+
+  map = new google.maps.Map(document.querySelector('#map'), mapOptions);
+  codeAddress();
+}
+
+const codeAddress = () => {
+  const address = restaurant.restaurant_address;
+
+  geocoder.geocode({ address: address }, function (results, status) {
+    if (status == 'OK') {
+      map.setCenter(results[0].geometry.location);
+      const marker = new google.maps.Marker({
+        map: map,
+        position: results[0].geometry.location,
+      });
+    } else {
+      alert('Geocode was not successful for the following reason: ' + status);
+    }
+  });
+};
